pyproject.toml
[tool.poetry]
name = "testing-pandas"
version = "0.1.0"
description = ""
authors = ["Gautam Singla <ggsingla777@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-generativeai = "^0.7.2"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pandas = "^2.2.2"
streamlit = "^1.37.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
